(maint) Support arbitrary URI schemes for remote targets in v2 inventory
In both the v1 and v2 inventory formats, we use the URI scheme as the
transport for the target. In v1 however, that was only true if the
target did not also separately have transport specified as "remote". In
that case, the URI scheme was used as simply a URI scheme (called
`protocol` when running a task) allowing it to therefore be arbitrary.
With the v2 inventory, we didn't respect that distinction and would a) fail
if the URI scheme did not correspond to a valid transport, and b)
override the "remote" transport specification with the one derived from
the scheme, preventing the use of remote targets with specific URI schemes.

We now vary the behavior of protocol and transport for v2 targets based
on whether or not they are remote. For remote targets, `transport` will
always be "remote" and `protocol` will be the URI scheme if specified.
For ordinary targets, the URI scheme will be used to set the `transport`
if specified and there will be no `protocol`.

On the Target datatype that is exposed to Puppet, we continue to
implement the `protocol` method for backward compatibility, returning
the value of `transport` for non-remote targets (which have no
protocol).