(maint) Support Kerberos authentication with SSH
The net-ssh-krb gem was previously licensed with an incompatible license
for use in Bolt, in addition to a dependency version conflict. The gem
was recently relicensed and updated, so can now be shipped out of the
box with Bolt and documented. This allows users to use Kerberos
authentication when connecting to targets over SSH by specifying
`gssapi-with-mic` as a `PreferredAuthenticationMethod` in their OpenSSH
config (typically `~/.ssh/config`).