(GH-1567) Use future_compile method as default in applicator
This replaces the previous functionality in `Bolt::Applicator#compile`
with the functionality in `Bolt::Applicator#future_compile`. When
compiling a catalog, the `Bolt::ApplyInventory` and `Bolt::ApplyTarget`
classes are now used instead of `Bolt::Inventory` and `Bolt::Target`.

This also updates `Bolt::Catalog#compile_catalog` to correctly override
target vars with plan vars.