(GH-1528) Provide a better error message when apply compilation fails
Previously, failed compilation of `apply()` blocks would result in a
stack trace with no useful information. The only way to determine what
had happened was to look back prior the failure at any logs that were
emitted during compilation.

We now cleanly handle failed compiles and deal with log messages
separately. The bolt_catalog command catches errors and returns them in
JSON form as the result of the command on stdout, along with exiting
non-zero to indicate a failure. This allows the applicator to cleanly
determine the proximate cause of the failure. We now omit the log
messages when compilation fails as they tend to duplicate the error that
was returned. We log them when compilation succeeds as they may include
important warnings in that case.

This also modifies the human outputter to handle ApplyResult and
RunFailure objects (which are both lists of Results) the same as a
ResultSet. This causes plans which fail due to an error to now have the
same output as if the error were caught and returned, rather than
printing large JSON blobs containing the errors.