(GH-1597) Add write_file plan function
This adds a new `write_file` plan function that writes content to a
specified file on the given targets. Internally, the function uses the
`file::write` function to write content to a tempfile on `localhost` and
then uses the `upload_file` function to upload the tempfile to the
destination on the given targets.