(GH-1612) Add --puppetfile CLI option
This adds support for a `--puppetfile` CLI option which accepts a
relative or absolute path to a Puppetfile to use when installing
modules with `bolt puppetfile install`. When using this flag, modules
will be installed to the current Boltdir. If a relative path is passed
to the option, it will be relative to the current Boltdir.