(GH-1623) Replace all 'nodes' keys with 'project migrate' command
This fixes a bug in the `project migrate` command that resulted in only
the first group in an array of groups having its `nodes` key replaced
with `targets`. The command now replaces `nodes` keys with `targets` in
*all* groups.

This also updates the `project migrate` command to remove any `version`
field that does not have a value of `2` from the inventory file.