(GH-1625) Load inventory more efficiently
The `resolve_string_targets` method was inefficient since it recomputed
the complete set of target names in the group for each target. We now
only compute the set once and keep it updated as we evaluate targets.
This saves ~0.5ms per target, which adds up when for users with
thousands or more targets.