(maint) Add contains_target? method in place of enumerating all targets
In several places in inventory, we use the `local_targets` method to
check whether a given target is contained by the group. That method
computes the list of targets every time, which is prohibitively
expensive for large inventories. We now use a `contains_target?` method
which checks for the presence of the target without making a list.