(maint) Don't copy entire config to merge transport data
Previously, we made a deep copy of and validated the entire Bolt::Config
object for every Target. That's quite expensive and creates a huge
amount of GC overhead, especially given that we're cloning complex
active objects like loggers and hitting the filesystem to expand paths
while validating. The only reason we did this copying was because the
logic that knew how to merge two transport configs together happened to
live on the Bolt::Config object and happened to make in-place
modifications to the config.

That method is now abstracted to a class method on Bolt::Config so it
can be used by the Target object without copying the entire config. We
now only make a clone of the transport config hash, which is a much
simpler ordeal.