(maint) Bump puppetlabs-reboot module
The latest puppetlabs-reboot version updates the deprecated `nodes`
parameter to `targets`. This is a breaking change for users who
explicitly set the parameter, either in the `run_plan` plan function or
on the CLI. Users will need to update the `nodes` parameter to be
`targets` instead. Calling the plan using `-t` or with
`run_plan('reboot', $mytargets)` will both behave the same and do not
need to be updated.