(maint) Warning instead of exceptiong whith invalid @param metadata
Previously if there was a descrepency between the @param descriptions in plan metadata and the actual parameters in the plan signature there would be an unhandled exception trying to extract parameter descriptions. With this commit if there is a @param description that describes a parameter that is not in the plan function signature a warning is raised and the @param metadata is ignored.