(GH-1591) Create transport config classes
This adds a `Bolt::Config::Transport` class and several derived classes
for each of Bolt's supported transports. These classes pull the logic
for filtering and validating config options for each of the transports
from the `Bolt::Config` and `Bolt::Transport` classes.

Each derived class is required to define `OPTIONS` and `DEFAULTS`
constants at a minimum.

The `OPTIONS` constant is a hash of valid option
keys mapped to a hash that defines the option's type and description.
Options set in the config file, inventory file, or via CLI options are
then filtered against the valid options keys and then have their values
validated against the option types.

The `DEFAULTS` constant is a hash of options mapped to the default
config options. When a transport config object is instantiated, the
`DEFAULTS` hash will be merged with any options passed to the
initializer.