(GH-1591) Update how transport config is used in Bolt
This modifies how transport config is passed between various Bolt
classes. The most significant change is how config is passed to the
`Bolt::Inventory` class. Previously, the entire `Bolt::Config` object
was passed to the inventory, which then accessed the necessary values.
Now, the inventory only receives the transport and transport
configuration from the config.

This also updates the inventory to handle resolving any plugin
references that are in the 'default' transport config. Since plugins are
now exposed under transport config settings in config files, the
transport config objects passed to the inventory may contain plugin
references. The inventory calls `@plugins.resolve_references` on each of
the transport config objects and updates the data with the
fully-resolved config.