(GH-1574) Initialize a Bolt project with modules
This adds support for initializing a Bolt project with a set of
specified modules and their dependencies. A new CLI option `--modules`
accepts a comma-separated list of module names (e.g. `puppetlabs-apt`).

Bolt uses the `puppetfile-resolver` gem to parse the list of modules
provided at the CLI and then resolves a dependency graph for the
modules. If any of the specified modules or their dependencies are
missing, the command errors and the project directory is not created. If
the dependency graph is resolved successfully, the list of modules and
dependencies are written to a Puppetfile in the new project directory
and `bolt puppetfile install` is immediately run against this
Puppetfile.

!feature

* **Initialize a Bolt project with modules and their dependencies**
([#1574](https://github.com/puppetlabs/bolt/issues/1574))

  The `bolt project init` command has a new `--modules` option that
  accepts a comma-separated list of modules to install when initializing
  a project. Modules and their dependencies are fully resolved, saved to
  a `Puppetfile` in the project directory, and then automatically
  installed with `bolt puppetfile install`.