(PE-28365) Make bolt-server compatible with bolt 2.0
With the old inventory support dropped in bolt 2.0 we need to port bolt-server to the new inventory. This commit simply requires bolt/inventory in the transport app to take advantage of the Bolt::Inventory.empty class method. This also fixes up the debugging script and docs to run bolt-server for 2.0 compatibility.