(maint) Include facts necessary for "warn" type
The provider for the warn type in the `basic` module used for testing `apply` is confined to agents with the `another` fact. Bolt spec tests currently only test that an apply block containing the warn type will compile. If the apply block is actually attempted to be run on an agent the warn type will not actually raise the warning because the `another` fact is not bundled in the plugin targball that is shipped to the agent. This commit includes the fact so that the warn type would raise the warning if applied to an agent.