(GH-1267) Support project level Puppet content
Users can now load puppet content from their Bolt project directory as
if it was a module. This adds the `name` configuration option to
`project.yaml` to specify a name for the project. If `project.yaml` is
present the project will be loaded as a module. The project is
namespaced either to the set `name` if present, or the name of the
project's directory. If the name conflicts with the boltlib modules the
project directory will not be loaded. This allows users to develop Bolt
content more easily by placing a `tasks` or `plans` directories directly
(ha) in their Bolt project and load.

Closes #1267

!feature

* **Experimental: Support project level Puppet content** ([#1267](#1267))

  Users can now load Puppet content from the root of the Bolt project directory, such as `<boltdir>/tasks`. Users must opt-in to this experimental feature by creating a `project.yaml` in their project directory.