(GH-1267) Add documentation for supporting top-level content
Now that we support loading the Bolt project directory as a module the
documentation around how modules are loaded needs to be updated. This
modifies docs that explicitly talk about the default modulepath and how
modules are loaded, but does not update examples that include the
`site-modules` directory as they are still valid.