(PE-28373) Always return a result-like error result
As part of the push to get orchestrator on 2.x it is expected that a bolt-server request will always expects a `status` and `value` key in the response. In most cases the Bolt::Result will have these keys. In the case where there is an error that is not target/action execution specific (in this case with a request schema validation error), return an error that has the `status` and `value` keys.

This commit also updates a stray `result` key to `value` in the scrub_stack_trace method.