Bolt::Util.deep_clone fails with frozen hashes
Our custom deep_clone method recurses on hashes and arrays to clone
their contents. This fails trying to clone frozen objects, since we can
modify the initial shallow clone of the object. This change saves the
frozen state of the original object, clones with `freeze: false` to
ensure the intermediate object is unfrozen, then sets the final object's
freeze bit once it's completely cloned. This ensures that the deep_clone
is identical to the original while still building it gradually with
recursion.

!bug

* **Bolt::Util.deep_clone can now clone frozen objects** ([#1696](https://github.com/puppetlabs/bolt/pulls/1696))

  The Bolt::Util.deep_clone method can now clone frozen objects, preserving the 'frozen' attribute