(maint) Remove outdated facts module from bolt puppetfile example
An example puppetfile used to demonstrate installing modules in a Bolt
project included an outdated version of the `puppetlabs-facts` module.
When users copy-paste the example the outdated module would override
versions that ship with Bolt and cause other tasks and plans (such as
puppet_agent::install) to fail. This removes the module from the example
Puppetfile.