(GH-813) Abstract Bash shell implementation from local+ssh transports
Previously, the similarities between the local and SSH transports were
represented by the Sudoable base transport class. This made the two
transports tightly coupled to bash. It also meant we had two separate
implementations of sudo support, one for each of the transports.

This commit introduces a conceptual split between the "connection"
portion of a transport and the "shell" that it talks to on the other
end. In this case, "local" and "SSH" are two implementations of how to
connect to a target, but the Bolt::Shell::Bash class contains all the logic of
how to actually get something done over the connection.

The SSH and local connection classes now are required to implement only
two methods: `execute` and `copy_file`. All other operations are built
in terms of those two primitives. All logic for how to run commands,
scripts, and tasks on a bash target, including how to invoke sudo, is
now contained in th Bolt::Shell::Bash class.

This represents the first step in eventually supporting running against
Powershell targets over SSH.