(GH-813) Pass environment variables directly to the task
Previously, we set environment variables with the `sudo` invocation and
relied on `sudo` to pass them through to the command being executed.
That doesn't work in every configuration or with every run-as command.
It also unnecessarily makes those values available to `sudo`. We now
set the environment variables *inside* the `sudo` invocation, ensuring
they're passed to the command we're trying to execute.

This changes the local transport to no longer rely on Open3's ability to
set environment variables directly. That means SSH and local now pass
environment variables in exactly the same way.

!bug

* **Don't rely on sudo to preserve environment variables**
([#1702](https://github.com/puppetlabs/bolt/pull/1702))

  Bolt previously set environment variables when invoking `sudo` and
  relied on it to preserve them when running the task executable. That
  behavior isn't reliable for all configurations, so now environment
  variables are set directly when running the underlying executable.