(GH-813) Don't fail if run-as user's homedir doesn't exist
When using run-as, we first `cd` to the user's homedir to ensure we're
not running from a directory that the run-as user can't access. However,
the run-as user's homedir doesn't necessarily have to exist. Previously,
that would cause a failure as the `cd` would fail and we would therefore
refuse to run the actual command. Now, we just *try* to `cd` to the
homedir if possible and then run the command regardless.

!bug

* **Don't fail when the `run-as` user's home directory doesn't exist**
([#1702](https://github.com/puppetlabs/bolt/pull/1702))

  When running commands with `run-as` set, Bolt will try to `cd` to the
  new user's home directory before running the command. If that fails
  because the directory doesn't exist, it will now run the command from
  wherever it currently is rather than aborting.