(GH-1713) Properly send UTF-8 characters over stdin
Previously, we were chunking the string to be sent to stdin into
4096-character chunks. When input is longer than the chunk size and
contains multi-byte characters, this causes incorrect behavior. We check
the return value of write_nonblock to determine how much to advance our
pointer to the stdin buffer. However, that method returns the number of
bytes written, not the number of characters. Therefore, when writing a
non-ASCII string, we will quickly become out of alignment and write
incorrect data to stdin.

We now make a binary copy of the input string and operate on it as a
byte buffer, ensuring that whatever we send will come out exactly the
same on the other side.

!bug

* **Fix a bug passing arguments to local shell transport** ([#1713](https://github.com/puppetlabs/bolt/1713))

  Local shell transport could miss some bytes when writing non-ASCII characters to stdin.