(GH-1722) Use 'targets' parameter in YAML plan steps
This deprecates the `target` parameter for YAML plan steps in favor of a
new `targets` parameter, in line with the 'magic' parameter used in
plans written in Puppet.

YAML plans that have steps with a `target` parameter will raise a
helpful warning notifying the user that the parameter is deprecated and
to use `targets` instead. YAML plan steps will still accept a `target`
parameter, but will prefer a `targets` parameter.

!deprecation

* **YAML plan step parameter `target` deprecated in favor of `targets`**
([#1722](https://github.com/puppetlabs/bolt/issues/1722))

  The `target` parameter for YAML plan steps has been deprecated in
  favor of `targets` and will be removed in a future release of Bolt.