(maint) Fix formatting and regex in changelog script
This makes two trivial changes to the changelog script:

- Add an extra newline after the last changelog entry. Previously,
headers were printed on the line immediately after the last entry.

- Use the commit message's raw body in `git log` pretty formatting.
Previously, only the commit message's body was being used. This would
cause any commit messages that had a release note label as the subject
of the commit to pass CI but be skipped in the changelog script.

!no-release-note