(GH-1725) Target facts should take precedence in apply
If a target has a fact that shares a name with a plan variable, a
target variable, or both, then the fact should take precedence when
passed in to apply. Previously, defining a fact with the same name as
another variable would cause a redefinition error in the apply block,
where now referencing the variable will refer to the fact value.

Closes #1725

!bug

* **Target facts with the same name as a plan or Target variable should
  not raise an error** ([#1725](https://github.com/puppetlabs/bolt/issues/1725))

  Previously, defining a fact with the same name as
  another variable would cause a redefinition error in the apply block,
  where now referencing the variable will refer to the fact value.