(GH-813) Wrap long powershell commands in a script for SSH
We can't execute commands that are 32700+ bytes long over SSH.
Therefore, we now wrap such commands in a script and invoke that. This
factors out the code that already does such wrapping for the local
transport and moves it to the Powershell class where it more rightly
belongs. This has the benefit that we no longer generate a script for
commands that are smaller than the limit when using the local transport.
Except for tasks like apply_catalog which receive large amounts of data
on stdin, most things should fit under the limit.