(maint) Periodically expire gem cache
Previously, the gem caches for jobs run through github actions were
based purely on a hash of the Gemfile and bolt.gemspec. This meant the
caches were only refreshed when we explicitly changed a dependency of
Bolt. Given that we do so relatively infrequently, that meant we could
miss updated versions of our gem dependencies. We now tie the gem cache
to the content of lib/bolt/version.rb, which changes every release. That
ensures we will update the gems we're testing against at least once a
week or so.

This also standardizes the way we handle installing bundler and caching
gems. The location we use for the bundler cache varied by job as did the
order in which we handled installing bundler and unpacking gems. We now
always follow the order a) install bundler, b) try to download the gem
cache, c) install gems if necessary.

!no-release-note