(GH-1673) Add run_task_with plan function
This adds a new plan function `run_task_with` that runs a task on a set
of targets with target-specific parameters. It accepts a lambda that
returns a `Hash` of parameters to pass to the task run on a particular
target.

As an example, a task might expect a parameter with a value that is set
by a target's `facts`. The `run_task_with` function would have a lambda
that assigns this parameter with the appropriate fact:

```
$results = run_task_with('my_task', $targets) |$t| {
    { 'osfamily' => $t.facts['osfamily'] }
}
```

Similar to the `run_task` plan function, `run_task_with` also accepts
an optional `description` parameter, as well as three metaparameters:

- `_catch_errors`
- `_noop`
- `_run_as`

!feature

* **Added `run_task_with` plan function**
([#1673](https://github.com/puppetlabs/bolt/issues/1673))

  The new plan function `run_task_with` lets you run tasks on a set of
  targets with target-specific parameters. It accepts a lambda that
  returns a `Hash` of parameters for a particular target.