(GH-1736) Support default task parameters for plugins
This adds support for passing default task parameters to plugins.
Previously, Bolt would merge parameters set in a `bolt.yaml` and
`inventory.yaml`, but not the default task parameters.

This also updates parameter validation to ignore any unrecognized
parameters set in a `bolt.yaml` for a specific plugin as long as other
plugins in the same module recognize the parameter.

!feature

* **Support default task parameters in plugins**
  ([#9999](https://github.com/puppetlabs/bolt/issues/9999))

  Bolt now merges default task parameters for a plugin with parameters
  set in a `bolt.yaml` and `inventory.yaml` file.