(GH-1749) Add post-2.0 contributions, scan for multiple release notes
This updates the changelog to include contribution lines for all
external contributions starting with Bolt 2.0.

This also updates the `parse_commit` method to scan a commit for
multiple release notes, adding them to the entries list separately.
Previously, this method would consider all text after the first matched
label as a release note, making it impossible to add multiple release
note in the same commit.