(PE-28877) Support bolt types in puppetserver compile endpoint
Allow the construction of an `ApplyTarget` with an empty config hash. This is useful in PE where the connection information is meaningless and is intentionally hidden from plan authors. Allow requiring a single file in puppetserver for constructing an ApplyInventory. By having all the dependencies loaded bolt-side we can just require a single file from the bolt library to construct an empty ApplyInventory for the compilation.