(GH-1773) Implement hash() method for Target
At a time in the past, Bolt::Target implemented both eql? and hash,
allowing Target objects to be uniqed and used as Hash keys. With the
switch over to inventory v2, Bolt::Target objects are generally considered
identical as long as they have the same name, so it makes even more
sense to be able to uniq them, etc. However, the current implementation
lacks the hash method that makes that behavior work.

We now implement hash() for both Bolt::Target and Bolt::ApplyTarget so
that plan authors can treat Target objects the same as target names.

!bug

* **Target objects of the same name are now identical** ([#1773](https://github.com/puppetlabs/bolt/issues/1773))

  Target objects will now be considered identical in all cases if they
  have the same name. This allows uniq to operate on arrays of Targets
  as well as Targets to be used as Hash keys.