(GH-1769) Don't fail when writing stdin to a task that has exited
Previously, when using run-as with stdin, both the local and SSH
transports would fail with a "broken pipe" error if the task exited
before all of stdin could be written. This happens if stdin is larger
than the pipe buffer and the task doesn't read stdin. We now anticipate
that error and simply stop trying to write stdin if it's encountered.

!bug

* **Fixed 'broken pipe' errors with SSH and local transports**
  ([#1769](https://github.com/puppetlabs/bolt/issues/1769))

  The SSH and local transports could experience broken pipes when
  using run-as while running a task that accepted input on stdin but
  didn't read it.