(GH-1736) Warn when using private-key or public-key options
This adds support for the `public-key` and `private-key` options for the
`pkcs7` plugin. Prior to modularizing the plugin, it accepted hyphenated
options. However, since the plugin is now a task, it can no longer
accept hyphenated options. Bolt will now automatically transform these
keys to `public_key` and `private_key` and issue a helpful warning
message that the hyphenated keys have been deprecated.

!deprecation

* **`private-key` and `public-key` options for `pkcs7` plugin have been
  deprecated**
  ([#1736](https://github.com/puppetlabs/bolt/issues/1736))

  The `pkcs7` plugin now accepts `private_key` and `public_key` options.
  Support for the `private-key` and `public-key` options will be removed
  in a future release of Bolt.