(GH-1759) Handle non-UTF-8 task output
Previously, we treated non-UTF-8 task output as simply non-JSON output,
putting it in a hash under the `_output` key. This violated a key
constraint of the task specification (task output must always be UTF-8)
and caused problems when we tried to print the result as JSON (either as
output or for logging).

We now explicitly check for non-UTF-8 characters when interpreting the
result and we throw away stdout entirely and replace it with an error if
it is invalid.

!bug

* **Task output that contains invalid UTF-8 is now rejected**
  ([#1759](https://github.com/puppetlabs/bolt/issues/1759))

  Tasks are defined as returning UTF-8, but Bolt didn't handle the
  non-UTF-8 case explicitly, leading to messy error messages and stack
  traces. The error should now be clear and meaningful.