(GH-1781) Add ResourceInstance data type
This adds a new `ResourceInstance` data type to the plan language and a
corresponding `Bolt::ResourceInstance` implementation class.

The `ResourceInstance` data type stores observed state, desired state  and
events for a single resource on a target. Resources are identified by a
combination of their target, type, and title, and can have their observed
state, desired state, and events set and overwritten.

This data type does not modify or interact with resources on a target in
any way.

!feature

* **Added `ResourceInstance` data type**
  ([#1781](https://github.com/puppetlabs/bolt/issues/1781))

  The new `ResourceInstance` data type is available for use in plans and
  can be used to store the observed state, desired state, and events for
  a target's resource.

!feature

* **Added `set_resources` plan function**
  ([#1781](https://github.com/puppetlabs/bolt/issues/1781))

  The `set_resources` plan function can be used to set
  `ResourceInstance`s on a `Target`.

!feature

* **Added `resources` function to `Target` data type**
  ([#1781](https://github.com/puppetlabs/bolt/issues/1781))

  `Target` objects have a new `resources` function that can be used to
  return a map of `ResourceInstance`s for the target.