(maint) Expand filepaths passed on the CLI relative to cwd
As part of the [#1364](https://github.com/puppetlabs/bolt/pull/1364) work to
standardize how we expand configured paths we expand all paths
configured in files relative to the project directory, and stated that
all paths set on the CLI would be expanded relative to the directory
Bolt was run from. However, only some CLI paths (inventoryfile and
puppetfile) were expanded this way, while others (hiera-config and
private-key) were expanded relative to the project
directory. This standardizes paths passed on the CLI to be expanded
relative to the cwd. This is accomplished by expanding the path in our
CLI option handler, so that the config object only sees the path as
absolute and does not try to expand it relative to the project.

!bug

* **Expand filepaths passed on the CLI relative to the current working directory**
  Config options hiera-config and private-key are now expanded relative
  to the directory Bolt was run from when specified on the CLI, inline
  with other CLI options.