(GH-1779) Warn when Bolt is installed as a gem
This adds a warning that Bolt may be installed as a gem when bundled
content is missing. Bolt checks if any modules other than `aggregate`,
`canary`, and `puppetdb_fact` are installed at `Bolt::PAL::MODULES_PATH`
to determine if it may be a gem install.

This warning can be disabled by setting the `BOLT_GEM` environment
variable to `false`.

!feature

* **Warn when Bolt is installed as a gem**
  ([#1779](https://github.com/puppetlabs/bolt/issues/1779))

  Bolt now issues a warning when it detects that it may have been
  installed as a gem. This warning can be disabled by setting the
  `BOLT_GEM` environment to `false`.

  To install Bolt reliably and with all of its dependencies, it should
  be [installed as a
  package](https://puppet.com/docs/bolt/latest/bolt_installing.html).