(maint) Fix plan show for project plans
`bolt plan show <plan>` was failing when passed a project-level
plan due to the fact that the project path is a Pathname rather than a
string, unlike most modules. We now convert the path to a string in
`get_plan_info` so that any code interacting with the plan data can
treat it as if it came from any other module.

!bug

  * **Fix `bolt plan show <plan>` for project-level plans**
    ([#1799](https://github.com/puppetlabs/bolt/pull/1799))

  This command was throwing errors due to a type mismatch that is now
  resolved.