(maint) Remove docker-compose download from linux CI
This removes the step to install docker-compose during pre-test setup
for the linux CI workflow. Previously, the linux virtual environment did
not have a version of docker-compose installed that supported the
`--parallel` flag. docker-compose has since been updated to a version
that does support this flag, so it's no longer necessary to download and
install docker-compose during pre-test setup.

!no-release-note