(GH-1771) Use plugins to set PuppetDB config
Previously, PuppetDB config was not able to be set using plugins. This
was due to the fact that PuppetDB config is used to configure the
PuppetDB plugin which can then be used to configure other plugins. We
already have logic that handles incrementally resolving data in the
`plugins` section of the config, but that couldn't accomodate PuppetDB
which was handled specially.

We used to construct the PuppetDB client and pass it to Bolt::Plugin during
plugin setup, where it would be used to construct the PuppetDB plugin
and then other plugins would be resolved thereafter. This commit flips
that dependency so that Bolt::Plugin takes the unresolved PuppetDB
config, merges it with the rest of the plugin config, and uses that to
construct the PuppetDB plugin. The CLI then takes a reference to the
PuppetDB client from the PuppetDB plugin, rather than the plugin taking
a reference to the CLI's client. This allows us to still have a single
PuppetDB client, but for that client to be configurable using plugins.

!feature

  * **Use plugins to set PuppetDB config** ([#1771](https://github.com/puppetlabs/bolt/issues/1795))

    Plugin references can now be used to set configuration options for
    the PuppetDB client used by Bolt, in the `puppetdb` section of the
    config.