(GH-1771) Expose puppetdb_client as Bolt::Plugin method
Previously, the PuppetDB client was accessed from the outside by
grabbing the PuppetDB plugin from the plugin system and taking its
client. We now hide that behavior inside a `puppetdb_client` method on
the Bolt::Plugin object so that consumers don't need to care about the
lifecycle of that client.