(maint) Add details to file function descriptions
The file module includes several helpful functions for interacting with
files locally. Previously the documentation didn't clearly specify that
the functions only act on disk and cannot be run against remote targets.
The function documentation now clarifies what exactly each function is
running "under the hood", and states that the functions are limited to
localhost.

!no-release-note