(PE-28674) No longer look for modulepath in bolt.yaml for PE plans
Attempting to use a modulepath from `bolt.yaml` in PE plans has exposed some issues that we are not ready to tackle yet. For example there is no interpolation for `$basemodulepath`. This commit prevents bolt-server from attempting to load a modulepath from bolt.yaml in an environment. This commit also updates `BoltServer::PE::PAL` to take advantage of the work done in the PAL we use in orchestrator to leverage threadsafe Puppet features.