(GH-1789) Rescue EMFILE error with helpful message
Since splitting the shell implementation from the transport connecting
to the shell, we now open a new thread with input, output, and error
streams for every connection. With concurrency defaulting to 100 this
often means 300+ file descriptors are open in Bolt's ruby process when
running against many targets, which exceeds the common FD limit of 256.
We now set concurrency to 1/3 the file descriptor limit if the limit is
below 300, otherwise it's set to 100. We also warn the user that
concurrency is set low, and provide instructions on how to increase the
ulimit or concurrency. We also raise a helpful error message when the
EMFILE error is thrown instructing the user on how to increase their FD
limit. Lastly this adds a section in the known issues documentation about
the error, also with instructions on increasing the FD limit and on
dialing back concurrency.

Closes #1789

!feature

* **Lower default concurrency when ulimit is low** ([1789](https://github.com/puppetlabs/bolt/issues/1789))

  Concurrency defaults to 1/3 the ulimit if ulimit is below 300, and
  warns if lowered concurrency is used.